% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggRuns.R
\name{aggRuns}
\alias{aggRuns}
\title{Aggregate Runs}
\usage{
aggRuns()
}
\description{
This funciton aggregates data from a series of runs/exeriments into single
data frame. Set the working directory to a folder containing the the runs
you wish to compile. The function searches for any files with 'net' in the
name, and compiles the 'NetShift' values.
}
