% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeBiosensorData.R
\name{analyzeBiosensorData}
\alias{analyzeBiosensorData}
\title{Process Raw Biosensor Data}
\usage{
analyzeBiosensorData(time1 = 51, time2 = 39,
  filename = "groupNames_allClusters.csv", loc = "plots", fsr = FALSE,
  chkRings = FALSE, plotData = TRUE, celebrate = TRUE, netShifts = TRUE,
  getLayoutFile = TRUE)
}
\arguments{
\item{time1}{a number specifying the later time for net shift calculations}

\item{time2}{a number specifying the earlier time for net shift calculations}

\item{filename}{a string with the filename containing the chip layout}

\item{loc}{a string with directory name to save plots and data files}

\item{fsr}{a logical value indicating whether the data contains FSR shifts}

\item{chkRings}{a logical value indicating if rings should be removed}

\item{plotData}{a logical value indicating if data should be plotted, which
will save a series of png files}

\item{celebrate}{a logical value, set it to TRUE for to be alerted when your
script has finished}

\item{netShifts}{a logical value indicating if net shift values should be
calculated and plotted}

\item{getLayoutFile}{a logical value indicating if the chip layout file
should be downloaded from Github}
}
\value{
This function returns csv files containing processed data along with
a number of png files containing plots of the processed data.
}
\description{
The purpose of this program is to process with the raw data from the
Maverick M1 detection system (Genalyte, Inc., San Diego, CA) and output
simple line graphs, bar charts, and box plots. In principle, this code should
also work for any bionsensor data that ouputs Time in column one and Signal
in column two. The function call also generates companion csv files
containning processed the prcoessed data for subsequent analysis.
The folder containing output from the M1 typically consists of:
   1. a csv file for each ring and
   2. a comments file the describes the experimental run
A "comments.csv" file is also exported, but it is not needed.
In addition to the csv files for each ring, a separate file containing the
chip layout is required. An example of a chip layout file is provided in the
"BaileyLabMRRs" repository located at
https://github.com/BaileyLabUM/BaileyLabMRRs. See the
"groupNames_allClusters.csv" file for an example.
}
\examples{
dir <- system.file("extdata", "20171112_gaskTestData_MRR",
                   package = "biosensor")
setwd(dir)
analyzeBiosensorData()

}
