% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrationStation.R
\name{calibrationStation}
\alias{calibrationStation}
\title{Calibration Station}
\usage{
calibrationStation(celebrate = TRUE)
}
\arguments{
\item{celebrate}{a logical value, set it to TRUE for to be alerted when your
script has finished}
}
\description{
Analyze series of runs on M1 and generate calibration curves for each target
}
