% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chopUpRun.R
\name{chopUpRun}
\alias{chopUpRun}
\title{Chop Up Run}
\usage{
chopUpRun(data, startRun, loc)
}
\arguments{
\item{data}{a dataframe containing at least two columns; the 1st is Time &
the second is Shift (signal from M1 Maverick Detection System)}

\item{startRun}{the numerical value on where to start the run, only used if
chopRun is TRUE}

\item{loc}{a string with directory name to save plots and data files}
}
\value{
Returns a data frame of the chopped up data, and it saves a csv file
of this data.
}
\description{
Cuts off all data before the time defines by $startRun variable. The the
response variables $Shift and $Time are set to zero at the $startRun point.
}
