% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelRuns.R
\name{labelRuns}
\alias{labelRuns}
\title{Label Runs}
\usage{
labelRuns(data, runsFile = "runs.csv")
}
\arguments{
\item{data}{a dataframe containing compiled experimental data generated from
the `aggRuns` function}

\item{runsFile}{a string with the filename of the runs file that contains the
experimental information for each run}
}
\description{
Annotate aggregated net shift data with run information
}
